- Set scss & routing - Desgin dashboard page
