+++ package.json
@@ -9,8 +9,13 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@solidjs/router": "^0.8.3",
+    "autoprefixer": "^10.4.16",
     "axios": "^1.5.1",
-    "solid-js": "^1.7.8"
+    "postcss": "^8.4.31",
+    "sass": "^1.69.3",
+    "solid-js": "^1.7.8",
+    "tailwindcss": "^3.3.3"
   },
   "devDependencies": {
     "vite": "^4.4.5",
